clk/qcom: introduce debugcc
debugcc is a tool for debugging the clock controllers on Qualcomm SoCs.
It uses the built-in clock debugging registers to measure clock rates by
muxing clocks to a special debug mux.

This is a port of the userspace tool [1], as it is plain C and only
depends on getopt, only minimal changes are necessary to integrate it
into u-boot. The main addition is a substring search to filter which
clocks are measured as this is quite useful when running on-device.

The "clk debug" command can be used to invoke "debugcc". This tool uses
the debug registers on the clock controller to measure the actual clock
frequencies.

List all core clock names:

  clk debug 0 -b core -l

Measure the usb30_prim_master clock:

  clk debug 0 gcc_usb30_prim_master_clk

Measure all clocks used by the SDcard controller:

  clk debug 0 -s sdcc2

| BLOCK    | CLOCK                | MHz     | Hz            |
+----------+----------------------+---------+---------------+
| core     | gcc_sdcc2_apps_clk   |  202MHz |   201501200Hz |
| core     | gcc_sdcc2_ahb_clk    |  100MHz |   100000860Hz |

Imported at commit b3f6a94 ("Update README")

[1]: https://github.com/linux-msm/debugcc

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>