dm: core: add support for stub drivers
Introduce support for a uclass to provide a fallback/stub driver which
can be used when no device is found for a given node. This might be
useful for stubbing non-vital clock controllers like the RPMh on
Qualcomm platforms, or during early bringup if the serial port has
already been configured.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>