clk: introduce a fallback stub driver
Add a stub_clk driver which does absolutely nothing and is configured as
the fallback stub for UCLASS_CLK. If there is a dependency on a clock
device which is not supported by any existing driver, and
CONFIG_CLK_STUB is enabled, then the stub driver will kick in.

This is intended to be useful during early bringup (e.g. if the serial
port is already configured for a default rate) or for clock controllers
that simply aren't vital for a platform to work.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>