qcom: RBx fixes and USB support
This series is a few loosely connected patches to get the RB1 and 2
boards booting from USB with upstream U-Boot, and a few preperatory
patches for rb5.

Unfortunately the RB5 board requires a regulator on the RPMh peripheral
to be turned on for the USB VBUS rail. Support for this will be added in
future patches.

* Add a few missing compatible strings for sm6115/sm8250
* Fix the femto-v2 HS USB phy used on sm8250
* Fix USB role selection on RB2 (without breaking Linux!)
* Pick a value for CONFIG_SYS_LOAD_ADDR that won't cause a null pointer
  on most boards
* Fix EFI booting by generating SMBIOS tables

To: Caleb Connolly <caleb.connolly@linaro.org>
To: Neil Armstrong <neil.armstrong@linaro.org>
To: Sumit Garg <sumit.garg@linaro.org>
To: Tom Rini <trini@konsulko.com>
Cc: u-boot-qcom@groups.io
Cc: u-boot@lists.denx.de

---
Changes in v2:
- Fix orientation of compatible check in ft_board_setup()
- Link to v1: https://lore.kernel.org/r/20240418-b4-qcom-rbx-fixes-v1-0-f5f19d52969e@linaro.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240418-b4-qcom-rbx-fixes-34676d77b83b",
    "prefixes": [],
    "prerequisites": [
      "base-commit: d5460b082cd6afd0e07c0ec0e5a82d1af8dc09f7"
    ],
    "history": {
      "v1": [
        "20240418-b4-qcom-rbx-fixes-v1-0-f5f19d52969e@linaro.org"
      ]
    }
  }
}