clk: add a stub clock driver
As U-Boot works to align itself with upstream devicetrees, there are
some common issues we start to run into, that of hardware blocks which
might be important for an OS like Linux, but which aren't useful in
U-Boot.

For example, the Resource Power Manager found on Qualcomm platforms
includes clock controllers and power domains which are only useful for
managing power consumption and enabling low power states.

As this is not at all relevant for U-Boot, we can safely ignore these
devices, but we don't have a good way to communicate that to U-Boot.

As an initial step, implement a "stub" clock driver which can be bound
to these devices so that drivers will work as expected. Add the
necessary compatibles for the rpmcc and rpmh-clk controllers on older
rpm (non-h) platforms and the sm8250/8550/8650 platforms.

This is necessary for certain peripherals to work on various platforms,
most notably the sdcard reader on SM8150 phones and the SM8250 RB5
development board.

To: Tom Rini <trini@konsulko.com>
To: Lukasz Majewski <lukma@denx.de>
To: Sean Anderson <seanga2@gmail.com>
To: Caleb Connolly <caleb.connolly@linaro.org>
To: Neil Armstrong <neil.armstrong@linaro.org>
To: Sumit Garg <sumit.garg@linaro.org>
Cc: Heinrich Schuchardt <xypron.glpk@gmx.de>
Cc: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Cc: u-boot@lists.denx.de
Cc: u-boot-qcom@groups.io

Changes in v4:
- Add sm8150 compatible
- Link to v3: https://lore.kernel.org/r/20240909-b4-clk-stub-v3-0-ef09307534d5@linaro.org

Changes in v3:
- Add 8250/855/8650 rpmh clk stubs
- Link to v2: https://lore.kernel.org/r/20240527-b4-clk-stub-v2-0-29013855e343@linaro.org

Changes in v2:
- Take the other discussed approach of just implementing a stub driver.
- Link to v1: https://lore.kernel.org/r/20240410-b4-stub-drivers-v1-0-6935bd2c07d1@linaro.org

To: Tom Rini <trini@konsulko.com>
To: Lukasz Majewski <lukma@denx.de>
To: Sean Anderson <seanga2@gmail.com>
To: Caleb Connolly <caleb.connolly@linaro.org>
To: Neil Armstrong <neil.armstrong@linaro.org>
To: Sumit Garg <sumit.garg@linaro.org>
Cc: u-boot@lists.denx.de
Cc: u-boot-qcom@groups.io

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 4,
    "change-id": "20240527-b4-clk-stub-9698797a65ef",
    "prefixes": [],
    "history": {
      "v1": [
        "20240410-b4-stub-drivers-v1-0-6935bd2c07d1@linaro.org"
      ],
      "v2": [
        "20240527-b4-clk-stub-v2-0-29013855e343@linaro.org"
      ],
      "v3": [
        "20240909-b4-clk-stub-v3-0-ef09307534d5@linaro.org"
      ]
    }
  }
}